Fixed up tests by only setting up hte database once, because Jasmine doesn't seem to do support async inside setup.

+++ test/spec/test_localstore.js
@@ -3,26 +3,38 @@
 define(['plasmid'], function(plasmid) {
   describe('Plasmid: LocalStore', function () {
 
-    var DB, ready = false;
-
-    // Initialize the controller and a mock scope
-    beforeEach(function () {
-      ready = false;
-    });
+    var DB
+    ,   ready = false
+    ,   names = [];
+    ;
 
     afterEach(function () {
-      if (!!DB.idb) {
-        indexedDB.deleteDatabase(DB.idb.name);
+      var closed = false;
+    
+      console.debug("closing...");
+      if (!!DB && !!DB.idb) {
+        var close = indexedDB.deleteDatabase(DB.idb.name);
+        close.onsuccess = function() {
+          console.debug("closed");
+        }
       }
     });
 
-    function make_database(schema) {
+    function make_database(name, schema) {
+      names.push(name);
       return function() {
+        console.log("Making database...", name, !!DB);
+        ready = false;
         DB = new plasmid.Database({
-          name: 'test',
+          name: name,
           schema: schema
         })
         .on('opensuccess', function() {
+          console.log("opened database");
+          ready = true;
+        })
+        .on('openerror', function() {
+          console.error("Could not open database");
           ready = true;
         })
       }
@@ -32,7 +44,7 @@ define(['plasmid'], function(plasmid) {
     }
 
     it('adds objects with unique keys', function () {
-      runs(make_database({
+      runs(make_database(1, {
         version: 1,
       }));
       waitsFor(wait_database);
@@ -63,6 +75,88 @@ define(['plasmid'], function(plasmid) {
       });
 
     })
+
+    var indexed_schema = {
+        version: 1,
+        stores: {
+            notes: {
+                sync: false,
+                indexes: {
+                    created: {key: 'created'}
+                }
+            }
+        }
+    };
+
+    it('walks over indices >', function () {
+      var one = 1; // new Date(2000, 0, 1);
+      var two = 2; // new Date(2000, 0, 2);
+
+      runs(make_database(2, indexed_schema));
+      waitsFor(function(){
+        console.log('db open?', ready);
+        return ready;    
+      }, "Database didn't open in time", 1000);
+
+      // create fixtures
+      var key = null;
+      runs(function(){
+        var notes = DB.stores.notes;
+        var promises = [];
+        promises.push(notes.add(null, {
+            created: one,
+            text: "one",
+        }));
+        promises.push(notes.add(null, {
+            created: two,
+            text: "two",
+        }));
+
+        plasmid.Promise.chain(promises)
+        .then(function(){
+          key = 2;
+        }).on('error', function(er){
+          console.log("Could not setup fixtures", er);    
+        });
+      });
+
+      waitsFor(function(){
+          return key !== null; }
+      , "fixtures to be created", 1000);
+
+      // query data
+      var done = false;
+      var length;
+      runs(function() {
+        DB.stores.notes.fetch({indexname: 'created', upto: two, exclusive: true})
+        .then(function(value) {
+          // expect on data
+          expect(value.length).toBe(1);
+          expect(value[0].value.text, "one");
+          done = true;
+        }, function(e) {
+          expect(true).toBe(false);
+          done = true;    
+        });
+      });
+
+      waitsFor(function(){
+          return done;
+      });
+      runs(function() {
+      });
+
+    })
+
+    runs(function(){
+      console.log("DONE");
+      for (var i=0; i < names.length; i++) {
+        var close = indexedDB.deleteDatabase(name[i]);
+        close.onsuccess = function() {
+          console.debug("closed");
+        }
+      }
+    });
 
   })
 })
